# Module5/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Calculator Project",
    "description": "My User Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Calculator",
      "description": "API for Calculus in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/calculator/add": {
      "get": {
        "tags": ["Calculator"],
        "summary": "Add 2 numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "description": "The First Number"
          },
          {
            "name": "num2",
            "in": "query",
            "description": "The Second number"
          }
        ],
        "responses": {
          "200": {
            "description": "This service allows you to add two numbers together"
          }
        }
      }
    },
    "/calculator/subtract": {
      "get": {
        "tags": ["Calculator"],
        "summary": "Subtract 2 numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "description": "First Number",
            "required": true,
            "type": "number"
          },
          {
            "name": "num2",
            "in": "query",
            "description": "Second Number",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": { "description": "Subtraction result" }
        }
      }
    },
    "/calculator/multiply": {
      "get": {
        "tags": ["Calculator"],
        "summary": "Multiply 2 numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "description": "First Number",
            "required": true,
            "type": "number"
          },
          {
            "name": "num2",
            "in": "query",
            "description": "Second Number",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": { "description": "Multiplication result" }
        }
      }
    },
    "/calculator/divide": {
      "get": {
        "tags": ["Calculator"],
        "summary": "Divide 2 numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "description": "First Number",
            "required": true,
            "type": "number"
          },
          {
            "name": "num2",
            "in": "query",
            "description": "Second Number",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": { "description": "Division result" },
          "400": { "description": "Division by zero error" }
        }
      }
    }
  }
}